+++ CHANGELOG.md
@@ -4,7 +4,7 @@ seleniumtid changelog
 v0.8.4
 ------
 
-*Release date: 2015-??-??*
+*Release date: 2015-03-05*
 
   * Allow to set firefox and chrome preferences from properties file  
     New config section [FirefoxPreferences] with firefox preferences, e.g. 'browser.download.dir = /tmp'